requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2==2.9.3
dj-rest-auth==2.2.1
django-allauth==0.41.0
